fixed linux compile error
<==== Mercurial data ====>
Branch: nat_traversal

+++ common/src/utils/network/http/server/http_server_connection.cpp
@@ -135,7 +135,7 @@ namespace nx_http
                     &msg.response->headers,
                     nx_http::HttpHeader(
                         "Content-Length",
-                        nx_http::StringType::number(contentLength.get()) ) );
+                        nx_http::StringType::number((qulonglong)contentLength.get()) ) );
         }
         else
         {
